Add write support for mirror file system.

+++ pywinfuse/fuseUnlink.py
@@ -2,6 +2,8 @@ import myWin32file
 import errno
 from tools import *
 import stat
+from logSys import *
+import ctypes
 
 class unlinkSupport:
   def unlink(self, path):
@@ -9,18 +11,25 @@ class unlinkSupport:
     return -errno.ENOSYS
   def rmdir(self, path):
     print '*** rmdir', path
-    return -errno.ENOENT
+    return -errno.ENOSYS
     
   def DeleteFileFunc(self, FileName, pInfo):# WINFUNCTYPE(c_int, LPCWSTR, PDOKAN_FILE_INFO))
     unixFilename = self.translateFileName(FileName)
-    st = self.getattr(unixFilename)
-    if st.st_mode & stat.S_IFDIR:
-      #Is DIR rmdir
-      return self.checkError(self.rmdir(unixFilename))
-    else:
-      print 'removing:',unixFilename
-      return self.checkError(self.unlink(unixFilename))
+    st = self.getattrWrapper(unixFilename)
+    if st != -errno.ENOENT:
+      if st.st_mode & stat.S_IFDIR:
+        #Is DIR rmdir
+        return self.checkError(self.rmdir(unixFilename))
+      else:
+        print 'removing:',unixFilename
+        return self.checkError(self.unlink(unixFilename))
+    return -myWin32file.ERROR_FILE_NOT_FOUND
 
+  def DeleteDirectoryFunc(self, FileName, pInfo): 
+    #dbg()
+    unixFilename = self.translateFileName(FileName)
+    cl('Filename:', unixFilename)
+    return self.checkError(self.rmdir(unixFilename))
 
 
 class writeSupport:
@@ -30,7 +39,7 @@ class writeSupport:
   def WriteFileFunc(self, FileName, Buffer, NumberOfBytesToWrite, NumberOfBytesWritten, Offset, pInfo): 
     #dbg()
     unixFilename = self.translateFileName(FileName)
-    realBuf = Buffer[:NumberOfBytesToWrite]
+    realBuf = ctypes.string_at(Buffer, NumberOfBytesToWrite)
     writtenLen = self.write(unixFilename, realBuf, Offset)
     if writtenLen < 0:
       return -myWin32file.ERROR_FILE_NOT_FOUND
